Start refactoring code into separate modules, got tests running in docker container.
